feat!: relax `Reflection`'s validation requirements
BEFORE:
A `Reflection` node is valid only if the *sum* of all its children sizes
is less than or equal to the size of the node.

AFTER:
A `Reflection` node is valid only if there's no child whose size is greater
than the size of the node.

BREAKING:
`data_tree::reflection::ConversionError::ExcessiveChildren` now has a
different set of fields.